+++ code/src/cljs/sixsq/nuvla/ui/ocre/views.cljs
@@ -5,7 +5,6 @@
   (:require
     ["papaparse" :as papa]
     [cljs.core.async :refer [<! timeout]]
-    [cljs.pprint :refer [cl-format pprint]]
     [clojure.spec.alpha :as s]
     [clojure.string :as str]
     [re-frame.core :refer [dispatch dispatch-sync subscribe]]
@@ -39,7 +38,7 @@
 (s/def ::code spec-utils/nonblank-string)
 (s/def ::activated ::spec-utils/timestamp)
 (s/def ::redeemed ::spec-utils/timestamp)
-(s/def ::target-audience spec-utils/nonblank-string)
+(s/def ::discipline spec-utils/nonblank-string)
 (s/def ::batch spec-utils/nonblank-string)
 (s/def ::wave spec-utils/nonblank-string)
 (s/def ::supplier spec-utils/nonblank-string)
@@ -48,6 +47,33 @@
 (s/def ::distributor str)
 (s/def ::platform spec-utils/nonblank-string)
 (s/def ::acl any?)
+(s/def ::institute spec-utils/nonblank-string)
+(s/def ::researcher-name spec-utils/nonblank-string)
+(s/def ::researcher-email spec-utils/nonblank-string)
+(s/def ::country #{
+                 "AF", "AX", "AL", "DZ", "AS", "AD", "AO", "AI", "AQ", "AG", "AR", "AM",
+                 "AW", "AU", "AT", "AZ", "BS", "BH", "BD", "BB", "BY", "BE", "BZ", "BJ",
+                 "BM", "BT", "BO", "BQ", "BA", "BW", "BV", "BR", "IO", "BN", "BG", "BF",
+                 "BI", "CV", "KH", "CM", "CA", "KY", "CF", "TD", "CL", "CN", "CX", "CC",
+                 "CO", "KM", "CG", "CD", "CK", "CR", "CI", "HR", "CU", "CW", "CY", "CZ",
+                 "DK", "DJ", "DM", "DO", "EC", "EG", "SV", "GQ", "ER", "EE", "SZ", "ET",
+                 "FK", "FO", "FJ", "FI", "FR", "GF", "PF", "TF", "GA", "GM", "GE", "DE",
+                 "GH", "GI", "GR", "GL", "GD", "GP", "GU", "GT", "GG", "GN", "GW", "GY",
+                 "HT", "HM", "VA", "HN", "HK", "HU", "IS", "IN", "ID", "IR", "IQ", "IE",
+                 "IM", "IL", "IT", "JM", "JP", "JE", "JO", "KZ", "KE", "KI", "KP", "KR",
+                 "KW", "KG", "LA", "LV", "LB", "LS", "LR", "LY", "LI", "LT", "LU", "MO",
+                 "MG", "MW", "MY", "MV", "ML", "MT", "MH", "MQ", "MR", "MU", "YT", "MX",
+                 "FM", "MD", "MC", "MN", "ME", "MS", "MA", "MZ", "MM", "NA", "NR", "NP",
+                 "NL", "NC", "NZ", "NI", "NE", "NG", "NU", "NF", "MK", "MP", "NO", "OM",
+                 "PK", "PW", "PS", "PA", "PG", "PY", "PE", "PH", "PN", "PL", "PT", "PR",
+                 "QA", "RE", "RO", "RU", "RW", "BL", "SH", "KN", "LC", "MF", "PM", "VC",
+                 "WS", "SM", "ST", "SA", "SN", "RS", "SC", "SL", "SG", "SX", "SK", "SI",
+                 "SB", "SO", "ZA", "GS", "SS", "ES", "LK", "SD", "SR", "SJ", "SE", "CH",
+                 "SY", "TW", "TJ", "TZ", "TH", "TL", "TG", "TK", "TO", "TT", "TN", "TR",
+                 "TM", "TC", "TV", "UG", "UA", "AE", "GB", "US", "UM", "UY", "UZ", "VU",
+                 "VE", "VN", "VG", "VI", "WF", "EH", "YE", "ZM", "ZW"
+                 })
+(s/def ::country-name spec-utils/nonblank-string)
 
 
 (s/def ::voucher
@@ -55,11 +81,16 @@
                       ::currency
                       ::state
                       ::code
-                      ::target-audience
-                      ::platform
+                      ::supplier
                       ::owner]
              :opt-un [::expiry
-                      ::supplier
+                      ::discipline
+                      ::institute
+                      ::researcher-name
+                      ::researcher-email
+                      ::country
+                      ::country-name
+                      ::platform
                       ::user
                       ::distributor
                       ::activated
@@ -69,7 +100,7 @@
                       ::batch
                       ::acl]))
 
-(def required-headers #{"code", "amount", "currency", "platform", "target-audience"})
+(def required-headers #{"code", "amount", "currency", "supplier"})
 
 (def show-modal (r/atom nil))
 
@@ -107,8 +138,9 @@
 (defn cimi-voucher
   [user-id json-line]
   (->> (select-keys json-line
-                    [:amount :currency :code :state :target-audience :supplier :expiry :platform
-                     :distributor :activated :service-info-url :redeemed :wave :batch])
+                    [:amount :currency :code :state :supplier :expiry :platform
+                     :distributor :activated :service-info-url :redeemed :wave :batch
+                     :discipline :institute :researcher-name :researcher-email :country :country-name])
        (remove #(nil? (second %)))
        (map transform-value)
        (into {})